+++ lib/src/commonMain/kotlin/com/seazon/feedme/lib/rss/service/gr/bo/GrItem.kt
@@ -60,11 +60,11 @@ data class GrItem(
             item.description = if (summary == null) null else (summary!!.content)
         }
         var audioUrl: String? = null
-        var audioSize = 0
+        var audioSize = 0L
         item.enclosure = enclosure?.map {
             if (audioUrl == null && it.type != null && (it.type!!.startsWith("audio/") || it.type!!.startsWith("text/"))) {
                 audioUrl = it.href
-                audioSize = it.length.toInt()
+                audioSize = it.length
             }
             RssEnclosure(
                 href = it.href,
